close the Header file unit after writing to it, otherwise, we get 0-length files on some platforms
Store dm in the namelist and use it to write out ALL components of the dx
array -- previously we were just doing the right thing in 2-d.

Still need to address why all processors write out the Header file -- I
think that we can safely switch this over to just the master processor.