allocating/deallocating gamma1_term within advance, initial_proj, and divu_iter.
note to Ann: make_S is responsible for setting gamma1_term.  At the beginning of advance_timestep, make_S is not called (since we extrapolate S to half time).  I manually set gamma1_term to zero here.  This will need to be revisited in the future if you ever want a non-zero gamma1_term.