This is a new sponge for only the xrb problem.  It's much easier to work with than the sponge that lives in MAESTRO.  In the other sponge, as you increase sponge_kappa, the sponge shape becomes very sharp to the point where it almost looks like a step function.
What I do is basically fit a cosine that goes from 1 to spongemin.  The sponging starts where the density is 10 times the anelastic cutoff and has a midpoint at the anelastic cutoff.  For now, you have to hard code in the value of spongemin into sponge.f90.  This can easily be changed later to be an inputs to probin.

As long as this file is in /xrb/, sponge_kappa has NO EFFECT on the problem.

According to John, spongemin cannot be equal to zero.  And the choice of a cosine for this sponge is certainly not final.