If flag=true (meaning we are computing the pert quantity in put_in_pert_form or X in convert_rhoX_to_X), I am using first order extrapolation to fill in non-periodic domain boundary ghost cells.
If flag=false, we are using the physical boundary conditions as defined in define_bc_tower.

Before, we were not computing the correct perturbational quantities in the non-periodic domain boundaries... in fact we weren't touching these ghost cells at all.  The benchmarks will need to be updated since the slope calculation depends on these ghost cells.