Removed the file add_thermal_to_force from the repository.
I now pass in "thermal" into mkrhohforce and add the contribution there.  I also had to add a flag as to whether or not to add thermal, since in the cell update step we don't add thermal conduction explicitly, but rather do the implicit solve.  For the predictor forces we do add thermal.

If thermal_diffusion=F, we set thermal=0.  Then, if predict_temp, we add the reaction terms to thermal.

So no functional changes here, just makes the code a lot prettier.