if enthalpy_pred_type .ne. 1, we use the gamma10 formulation in paper_ann.tex.
I have tested this with enthalpy_pred_type .eq. 1 and there results are
identical, as expected.

We should run the paper III tests with the other enthalpy_pred_types to make
sure I didn't mess anything up.

The next thing that needs to be done with the code is make sure (rho h)_0 and
(rho X)_0 are never updated or even accessed if enthalpy_pred_type .ne. 1
... a good sanity check here is that the results for enthalpy_pred_type .ne. 1
should be identical before and after the next set of changes.