add a new flag, quadratic_interpolation, which turns on/off the quadratic interpolation of the initial model onto the base state.  The default is true (this is the old behavior).
Quadratic interpolation can screw up the HSE-ness of the initial model,
even if the initial model and base state have the same dr.

Also add a new diagnostic that prints out the maximum relative error
in the HSE-ness of the initial model, after it is mapped onto the base
state arrays.