Fancy new way for doing multilevel initialization of s0_init and p0_init.  For each level, I load in a different model file corresponding to the resolution of the level.  I've hardwired probin and varden to assume that we have at most 4 total levels since I couldn't figure out how to allocate arrays on the fly in probin.  This can be easily changed in the unlikely event that we do problems with more than 4 levels, or if somebody else is clever enough to figure out how to allocate arrays within probin.
I had to use a special trick to deal with the anelastic cutoff.  We initialize the coarsest level first.  Once we pass the cutoff, instead of storing the cell-centered location of the cutoff, I just store the values of all variables above the cutoff.  Doing it the old way caused a mismatch in the values of the variables above the cutoff since the interpolation was being used on different models.  This way, the numerical values are exact and consistent above the cutoff for all levels.

The multilevel inputs files now need to point to each a model file for each level, done as:

model_file = ""
model_file_lev2 = ""
model_file_lev3 = ""
... etc

I still haven't used the restrict_base operator in varden yet to synchronize the levels.  This is coming soon...