pass both rho0_old and rho0_new into make_w0.  In the first half of the algorithm, they will both be set to rho0_old.  This allows for time-centering for the spherical make_w0.
Slight changes may be present for planar problems, since rho0 was being
used to compute the base_cutoff, but now I switch to base_cutoff_density_coord.
Previously, the first half of the algorithm was using rho0_old and the second
half was using rho0_new to compare to base_cutoff_density, which was not right.