Marged all the firstdt.f90 files into one common file in /MAESTRO/ without changing any of the results.
There are two new input flags in probin "use_soundspeed_firstdt" and "use_divu_firstdt".  They both default to false.  For the xrb, wdconvect, and spherical_heat problems, I had to make them TRUE in the inputs file to preserve functionality.

NOTE that the divu constraint for spherical is still not implemented.