new probin option "full_rhoh0_evolution", currently false so no change in results.
I enabled True for the xrb problem.

What this does is average rho h after react_state and store that result in rhoh0_1.  Advect base evolves this to rhoh0_2.  Then rhoh0 is not used any more for the rest of the time step, so rhoh0_1 is set again during the next time step