Bug fix -
I added a call to fill_ghost_base at the end of average for the non-spherical, fine_grids_span_domain_width=T case.

Some quantities when averaged, such as gamma1bar and rhoh, need ghost cells filled on the 1d array.  Not all of them do, but the call to average is super cheap and this way we don't have to think about it in the future.