Addressed the following lines in TODO:
-- make firstdt do the divu estimator correctly for spherical

-- use the new w0mac stuff for estdt and makevort

The timestep for spherical problems may slightly change since the divU constraint is coded properly in firstdt, and the interpolation type of w0 to w0mac has been improved to be consistent with the rest of the code.  Using inputs_3d.test, this has no effect.

radial_velocity and velplusw0 in the plotfile are now slightly difference since we are using a new interpolation type for w0.  It actually looks a bit smoother, particularly near the corners of the domain.