instead of declaring 1d arrays as allocatable and then allocating them, you can just declare them as the proper size directly.
also, no need to deallocate these arrays since that is done automatically as you exit the function.  The only thing that needs a deallocate statement is a pointer that has been allocated, or a built multifab (via the destroy command).

pretty much every other file in MAESTRO could go through these changes, but this will have to wait.