start of the changes to return rho_Hnuc from the burners.
At the moment, we create a rho_Hnuc2 in varden/initialize, and carry it
through to all the same places that rho_omegadot2 is carried.

rho_Hnuc2 is now stored in checkpoints, so you won't be able to restart
from an old checkpoint file with this new code.

rho_Hnuc2 is not actually used anywhere (and make_plotfile will eventually
need to be changed to store it rather than recomputing it from the
binding energies).