more interface changes to support creating rho_Hnuc in the burners and passing it around as needed.
Note: rho_Hnuc is not actually used anywhere yet, but all of the
interface changes to the main Maestro routines are done.

To make rho_Hnuc actually be used, we still need to update
add_react_to_thermal.f90, make_S.f90, react_base.f90, and react_state.f90,
as well as change the interfaces to the extern/network routines.