Made a noteworthy notational change:
\newcommand{\uadvone}{\widetilde{\mathbf{U}}^{\mathrm{ADV},(1)}}
\newcommand{\uadvtwo}{\widetilde{\mathbf{U}}^{\mathrm{ADV},(2)}}

has now become

\newcommand{\uadvone}{\widetilde{\mathbf{U}}^{\mathrm{ADV},\star}}
\newcommand{\uadvtwo}{\widetilde{\mathbf{U}}^{\mathrm{ADV}}}

I really think this change has to be made because

the ^{n}, ^{(1)}, ^{(2)}, ^{n+1} notation refers to different points in the splitting scheme.

The ^{\star} vs ^{no star} notation has to do with the predictor/corrector algorithm.

The old definitions of \uadveone and \uadvtwo was mixing this ideas, and it was notationally inconsistent.