an adaptive algorithm fix -
after regridding, we need to recompute h = h(rho,p0,X).  Otherwise, the velocity field starts to checkerboard near the cutoff density.  This fix eliminates the problem for the first few thousand time steps - I'll have to run this for a full day to see if it's the permanent solution.