break out the diag's for xrb
note that T_max is not that useful as is because the base of the domain is
the hottest region and it takes a long time for the burning region to reach
this temperature.  It might be useful to tweak the calculation of T_max to only
be calculated within a small range centered about the helium transition layer
but this would be problem dependent.