This was my first stab at some analysis routines for VisIt.  It works but is very cumbersome, slow for large data-sets, only allows for BoxLib data and has a very inefficient way to specify plot attributes such as limits, logscale, etc.
That being said, I'm currently rewritting the script to attack some of these
issues, but I wanted to have a backup copy somehwere so that I can still do
analysis.

This script assumes that VisIt is installed and that ${PYTHONPATH} knows where
to look for the VisIt python interface libraries and modules.  I've noticed
on some machines (Davinci, for example), finding the locations of said
libraries is almost impossible because of the custom install and the fact that
the filenames of the required libraries/modules can be different from those on
other systems depending on the version of VisIt installed.  However, VisIt's
python interface can be run through a VisIt-provided python interpretter called
the Command Line Interface by calling VisIt with the "-cli" option.  This
script will run in that case if all instances of "import visit" and all object
references "visit." are removed; i.e. change "visit.Launch()" to just
"Launch()".