added a mass fraction cutoff parameter for the burning loop
the species to check for cutoff is designated by the
burner_threshold_species string which should be a name or short name
of a species in the network.  if that name is not found in the
network, no threshold condition is not checked and burning proceeds
normally.  this string defaults to ""

the mass fraction cutoff for burner_threshold_species is specified by
the burner_threshold_cutoff variable.  it's default is set to 1.e-10.

burning will only occur in those zones where
x_in(burner_threshold_species) > burner_threshold_cutoff

this may help cut down some unnessecary calls to the burner where
there should not be much burning going on