add the ability to do buffering to the diagnostic output.  A new routine, flush_diag() is now called just before each checkpoint file is written out.  To buffer, a problem's diag.f90 should allocate storage for many timestep's worth of diagnostic data, and store it, rather than write it out each timestep.  Then, whenever flush_diag() is called, the data is written to disk all at once.  A new runtime parameter, diag_buf_size can be used to force writing to disk at regular intervals.  Currently, only the wdconvect version of diag.f90 does buffering.
On machines with slow I/O, this should help speed things up.

Also, fixed a formatting error in the wdconvect diag output -- 11 columns
of data were being written to the enuc and temp diag files, but the format
statement only had 10 entries, so the 11th appeared on the next line.
This is more annoying than a bug, and gnuplot seems to ignore this.  A
simple script can repair any existing output.