add restart_with_vel_field flag which allows for a restart of the problem using the velocity field from a drive_initial_convection = T run.
note this is not currently user friendly.  To restart in this fashion:

  1) set restart_with_vel_field = T

  2) set restart = the checkfile number (from the
                   drive_initial_convection = T run) you want to restart from

Upon restart, the velocity field information is loaded into the state
fabs and all other variables are loaded in the init_base_state and
initscalardata routines.  The time is reset to ZERO.  The "restart"
variable is set to -1 to force the normal initialization iters.  This
means that the chkfile you restarted from will likely be overwritten
unless you change chk_int from what it was in the
drive_initial_convection run.