if we are running with evolve_base_state = F and multilevel, then when we regrid, we need to make sure that the base state data is valid in any newly created fine-grid regions.
Now, if evolve_base_state = F, we use piecewise linear interpolation from
the coarse level to fill in the new regions of the rho0 array.  This
mimics the procedure used for eta_cc and psi: interpolate from the coarse
grid into all the finer cells and then copy over the existing rho0 data.
Next we regrid and then zero out rho0 in regions where there is a finer
overlying grid.

Also, we now call enforce_HSE for both evolve_base_state = T or F to get
a valid p0 in the new regions.