added flag do_heating which allows for turning on and off of the heating
now, we only go into the get_rho_Hext and/or burner_loop routines if
do_heating and/or do_burning are T.  This bypasses having to all the
way down to the dataptr level and setting things ZERO when we wanted
to turn something off.  Note this introduces order of operation
differences in some quantities, so this might break some regression
testing.

this fix has been on my backburner (no pun intended) for some time