added some f2py directives (in comments)
changed a parameter definition, which had (un-needed) extra
parenthesis that just confused f2py.  the value of the parameter did
not change.

commented out the "use parallel" and the diagnostic print statements
for reading the helm_table.  this simplifies things in the f2py'ing
and lets the /data_processing/python/helmeos.py module work properly.
If this causes problems, let me know and I'll change the way I'm doing
the f2py'ing and revert this part of the change to helmeos.f90