put the default BoxLib define_bc_tower.f90 here to avoid picking up the MAESTRO version.  This is needed because this unit test has a custom variables.f90 and doesn't define all the variables that MAESTRO usually wants.
Unfortunately, there is an issue with files with the same name living
in both BoxLib and MAESTRO.  The way things are setup now, we pick up
the MAESTRO versions -- this is probably what we want, since these
are MAESTRO unit tests.  But in this instance, we don't want the MAESTRO
version and there is not an easy way at the moment to tell the build
system this.