change how we look for the inputs file.
No longer look at the PROBIN environment variable (don't think
that anyone was using that), or for a default inputs_maestro.

Instead, now we always require the inputs file to be specified
as the first commandline argument.  We abort if one is not
specified.  We also abort with a message if the file is not
found.  This greatly clarifies the behavior of the code.