regrid() takes an auxillary mf that can be used in the tagging.  By default, we are passing in either rho_Hnuc or rho_Hext (depending on whether or not we are burning).  But sometimes it makes sense to look at tpert.  If we set use_tpert_in_tagging = T, now tpert will be passed in instead.
Eventually we may want to generalize things to allow more stuff to
be passed in (perhaps passing tag_mf() off to a separate subroutine
that can be problem-dependent.  For now this should incur no more
expense over the default, and this will allow us to merge the old
(and currently broken) test2_adaptive stuff into the main test2 problem.