Fix segfault triggered by Cray cce >= 8.4.0.
On Titan, segfaults were occuring in make_pi_cc.f90:make_pi_cc_*d() for
Cray compilers cce >= 8.4.0.  The error does not occur for PGI, so it seems to
be a compiler bug.  However, replacing use of the incoming optional argument
with a local array fixes the problem.  I'd prefer not to change code to
work around compiler bugs, but in this case the fix is fairly easy and
only requires the slight cost of now having a local automatic array.