Add particles to sub-Ch.
For the purposes of verifying hot fluid initially below the convection gets
advected eventually and initial post-processing functionality, I have added
particles to thr sub-Chandra build.  In init_particles you can specify a number
of rays and number of particles per ray.  A number of rays will be uniformy
distributed across angles and then for each ray the number of particles you
specify will be uniformly distributed across the convecting region, starting a
bit below it and ending at the anelastic cutoff (near the top of/above the
convecting region).