Make burning_cutoff_density a user-controllable parameter.
Previously, burning_cutoff_density was set to the base_cutoff_density
at initialization and the user could not change burning_cutoff_density
at runtime.

This initializes burning_cutoff_density to -1.0d0 and if the user has
not set it to a positive value, then set it equal to
base_cutoff_density. This does not therefore require any changes to
existing inputs files.